refactor(modal): 모달의 디렉토리 위치 변경
