+++ components/src/lib/storybook/Modal.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import Modal from "../Modal";
+import Modal from "../Modal/common/Modal";
 
 const meta = {
   title: "ModalExample",
